Setter Injection
Now let's go a step further... In SlackClient, I want to log a message. But, we already know how to do this: add a second constructor argument, type-hint it with LoggerInterface and, we're done!

But... there's another way to autowire your dependencies: setter injection. Ok, it's just a fancy-sounding word for a simple concept. Setter injection is less common than passing things through the constructor, but sometimes it makes sense for optional dependencies - like a logger. What I mean is, if a logger was not passed to this class, we could still write our code so that it works. It's not required like the Slack client.

Anyways, here's how setter injection works: create a public function setLogger() with the normal LoggerInterface {dollar_sign}logger argument. Create the property for this: there's no shortcut to help us this time. Inside, say {dollar_sign}this->logger = {dollar_sign}logger. In sendMessage(), let's use it! Start with if ({dollar_sign}this->logger). And inside, {dollar_sign}this->logger->info(). Bah! No auto-complete: with setter injection, we need to help PhpStorm by adding some PHPDoc in the property: it will be LoggerInterface or - in theory - null. Now it auto-completes ->info(). Say, 'Beaming a message to Slack! '

In practice, the if statement isn't needed: when we're done, Symfony will pass us the logger, always. But... I'm coding defensively because, from the perspective of this class, there's no guarantee that whoever is using it will call setLogger().

So... is Symfony smart enough to call this method automatically? Let's find out - refresh! Our class still works... but check out the profiler and go to 'Logs'. Bah! Nothing is logged yet!