The @required Directive
Yep, Symfony's autowiring is not that magic - and that's on purpose: it only autowires the __construct() method. But... it would be pretty cool if we could somehow say:

    Hey container! How are you? Oh, I'm wonderful - thanks for asking. Anyways, after you instantiate SlackClient, could you also call setLogger()?

And... yeah! That's not only possible, it's easy. Above setLogger(), add /** to create PHPDoc. You can keep or delete the @param stuff - that's only documentation. But here's the magic: add @required

As soon as you put @required above a method, Symfony will call that method before giving us the object. And thanks to autowiring, it will pass the the logger service to the argument.

Ok, move over and... try it! There's the Slack message. And... in the logs... yes! We are logging!