For our last trick, I want to introduce a bundle that's going to make our life awesome. And, for the first time, we are going to hook into Symfony. Installing Maker Bundle
First, find your terminal, and install that bundle:
composer require maker --dev

Yep! That's a Flex alias for symfony/maker-bundle. And, in this case, 'make' means - 'make your life easier by generating code'.

We know that the main purpose of a bundle is to give us services. And, that's true in this case too... but the purpose of these services isn't for us to use them directly, like in our controller. Nope, the purpose of these services is that they give us new bin/console commands:

php bin/console

Nice! About 10 new commands, capable of generating all kinds of things. And, more make commands are still being added.