Let's make our command a bit more fun! Give it a description: 'Returns some article stats'
Each command can have arguments - which are strings passed after the command and options, which are prefixed with --, like --option1:

php bin/console article:stats arg1 arg2 --option1 --opt2=khan

Rename the argument to slug, change it to InputArgument::REQUIRED - which means that you must pass this argument to the command, and give it a description: 'The article's slug'. Rename the option to format: I want to be able to say --format=json to get the article stats as JSON. Change this to VALUE_REQUIRED: instead of just --format, this means we need to say --format=something. Update its description, and give it a default value: text.

Perfect! We're not using these options yet, but we can already go back and run the command with a --help flag:

php bin/console article:stats --help

Actually, you can add --help to any command to get all the info about it - like the description, arguments and options... including a bunch of options that apply to all commands.