Customizing our Command
Ok, so the configure() method is where we set things up. But execute() is where the magic happens. We can do whatever we want here!

To get the argument value, update the getArgument() call to slug and rename the variable too.Let's just invent some article 'data', give this array a slug key and, how about, hearts set to a random number between 10 and 100. Clear out the rest of the code, and then add a switch statement on {dollar_sign}input->getOption('format'). Here's the plan: we're going to support two different formats: text - don't forget the break - and json

If someone tries to use a different format, yell at them!

    What kind of crazy format is that?